refactor(utils:gpu): rename gpu_adapter to gpu
